drivers: jesd204-core: add basic caps & validation
This change adds support for specifying the JESD204 revision and subclass
that the device supports and should operate at.
These will be specified via DT, and will be read and validated whenever the
jesd204_dev_register() function gets called.
They've been added to a `dt-bindings/jesd204/jesd204.h` header.

It could (potentially) happen that a device could support more than 1
JESD204 revisions and/or subclasses, but it's not likely that a dynamic
re-configuration (at this level) would be desired (ever).

These parameters will be used to validate the requirement of SYSREF and
SYNC~ signals (based on subclass type & revision type). As well as other
parameters that may change later with newer JESD204 revisions (if any).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>