drivers: jesd204-core: add topology creation logic via OF graph
This change is probably the bread-n-butter of this driver as it allows the
discovery of devices/elements required to create a JESD204 link.

This is done by using OF graph to iterate over other buses that may
contain devices (clocks, converters, transceivers, peripherals, etc).
So far, it scans the platform, I2C, SPI & IIO buses.
This is where typical devices (that are part in a JESD204 link) could
reside.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>